+++ test/Ether.Network.Tests/Context/MyClient.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Ether.Network.Packets;
 
-namespace Ether.Network.Tests.Server
+namespace Ether.Network.Tests.Context
 {
     internal class MyClient : NetClient
     {
